Moved get_optic and _calculate_facet_pointing methods to abstract process sofast class.
