+++ opencsp/app/sofast/lib/ProcessSofastFixed.py
@@ -1,14 +1,11 @@
 """Class that handles the processing of fixed pattern deflectometry data.
 """
 
-from typing import Literal
-
 import cv2 as cv
 import numpy as np
 from numpy import ndarray
 
 from opencsp.app.sofast.lib.BlobIndex import BlobIndex
-import opencsp.app.sofast.lib.calculation_data_classes as cdc
 from opencsp.app.sofast.lib.DefinitionEnsemble import DefinitionEnsemble
 from opencsp.app.sofast.lib.DefinitionFacet import DefinitionFacet
 from opencsp.app.sofast.lib.DotLocationsFixedPattern import DotLocationsFixedPattern
@@ -19,15 +16,8 @@ from opencsp.app.sofast.lib.ProcessSofastAbstract import ProcessSofastAbstract
 import opencsp.app.sofast.lib.process_optics_geometry as pr
 from opencsp.app.sofast.lib.SpatialOrientation import SpatialOrientation
 from opencsp.common.lib.camera.Camera import Camera
-from opencsp.common.lib.csp.Facet import Facet
-from opencsp.common.lib.csp.FacetEnsemble import FacetEnsemble
-from opencsp.common.lib.csp.MirrorPoint import MirrorPoint
 from opencsp.common.lib.deflectometry.SlopeSolver import SlopeSolver
 from opencsp.common.lib.deflectometry.Surface2DAbstract import Surface2DAbstract
-from opencsp.common.lib.geometry.RegionXY import RegionXY
-from opencsp.common.lib.geometry.TransformXYZ import TransformXYZ
-from opencsp.common.lib.geometry.Uxyz import Uxyz
-from opencsp.common.lib.geometry.Vxyz import Vxyz
 from opencsp.common.lib.geometry.Vxy import Vxy
 import opencsp.common.lib.tool.log_tools as lt
 
@@ -369,116 +359,3 @@ class ProcessSofastFixed(ProcessSofastAbstract):
 
         # Calculate facet pointing
         self._calculate_facet_pointing()
-
-    def _calculate_facet_pointing(self, reference: Literal['average'] | int = 'average') -> None:
-        """
-        Calculates facet pointing relative to the given reference.
-
-        Parameters
-        ----------
-        reference : 'average' | int
-            If 'average', the pointing reference is the average of all
-            facet pointing directions. If, int, that facet index is assumed
-            to have perfect pointing.
-        """
-        if self.data_calculation_facet is None:
-            lt.error_and_raise(ValueError, 'Slopes must be solved first by running "solve_slopes".')
-        if (reference != 'average') and not isinstance(reference, int):
-            lt.error_and_raise(ValueError, 'Given reference must be int or "average".')
-        if isinstance(reference, int) and (reference >= self.num_facets):
-            lt.error_and_raise(
-                ValueError, f'Given facet index, {reference:d}, is out of range of 0-{self.num_facets - 1:d}.'
-            )
-
-        # Instantiate data list
-        self.data_calculation_ensemble = []
-
-        trans_facet_ensemble_list = []
-        v_pointing_matrix = np.zeros((3, self.num_facets))
-        for idx in range(self.num_facets):
-            # Get transformation from user-input and slope solving
-            trans_1 = TransformXYZ.from_R_V(
-                self.data_ensemble_def.r_facet_ensemble[idx], self.data_ensemble_def.v_facet_locations[idx]
-            )
-            trans_2 = self.data_calculation_facet[idx].trans_alignment
-            # Calculate inverse of slope solving transform
-            trans_2 = TransformXYZ.from_V(-trans_2.V) * TransformXYZ.from_R(trans_2.R.inv())
-            # Create local to global transformation
-            trans_facet_ensemble_list.append(trans_2 * trans_1)
-
-            # Calculate pointing vector in ensemble coordinates
-            v_pointing = Vxyz((0, 0, 1)).rotate(trans_facet_ensemble_list[idx].R)
-            v_pointing_matrix[:, idx] = v_pointing.data.squeeze()
-
-        # Calculate reference pointing direction
-        if isinstance(reference, int):
-            v_pointing_ref = Vxyz(v_pointing_matrix[:, reference])
-        elif reference == 'average':
-            v_pointing_ref = Vxyz(v_pointing_matrix.mean(1))
-        # Calculate rotation to align pointing vectors
-        r_align_pointing = v_pointing_ref.align_to(Vxyz((0, 0, 1)))
-        trans_align_pointing = TransformXYZ.from_R(r_align_pointing)
-
-        # Apply alignment rotation to total transformation
-        trans_facet_ensemble_list = [trans_align_pointing * t for t in trans_facet_ensemble_list]
-
-        # Calculate global slope and surface points
-        for idx in range(self.num_facets):
-            # Get slope data
-            slopes = self.data_calculation_facet[idx].slopes_facet_xy  # facet coordinats
-
-            # Calculate surface normals in local (facet) coordinates
-            u_surf_norms = np.ones((3, slopes.shape[1]))
-            u_surf_norms[:2] = -slopes
-            u_surf_norms = Uxyz(u_surf_norms).as_Vxyz()
-
-            # Apply rotation to normal vectors
-            u_surf_norms_global = u_surf_norms.rotate(trans_facet_ensemble_list[idx].R)
-            # Convert normal vectors to global (ensemble) slopes
-            slopes_ensemble_xy = -u_surf_norms_global.data[:2] / u_surf_norms_global.data[2:]
-
-            # Convert surface points to global (ensemble) coordinates
-            v_surf_points_ensemble = trans_facet_ensemble_list[idx].apply(
-                self.data_calculation_facet[idx].v_surf_points_facet
-            )
-
-            # Calculate pointing vectors in ensemble coordinates
-            v_facet_pointing_ensemble = Vxyz((0, 0, 1)).rotate(trans_facet_ensemble_list[idx].R)
-
-            data = cdc.CalculationFacetEnsemble(
-                trans_facet_ensemble_list[idx], slopes_ensemble_xy, v_surf_points_ensemble, v_facet_pointing_ensemble
-            )
-            self.data_calculation_ensemble.append(data)
-
-    def get_optic(
-        self, interpolation_type: Literal['given', 'bilinear', 'clough_tocher', 'nearest'] = 'nearest'
-    ) -> Facet | FacetEnsemble:
-        """Returns mirror object with slope data"""
-        facets = []
-        trans_list = []
-        for idx_facet in range(self.num_facets):
-            # Get mirror surface points
-            v_surf_pts = self.data_calculation_facet[idx_facet].v_surf_points_facet
-            # Get point normal vectors
-            v_normals_data = np.ones((3, len(v_surf_pts)))
-            v_normals_data[:2, :] = self.data_calculation_facet[idx_facet].slopes_facet_xy
-            v_normals_data[:2, :] *= -1
-            v_normals = Uxyz(v_normals_data)
-            # Get optic shape
-            shape = RegionXY.from_vertices(self.data_facet_def[idx_facet].v_facet_corners.projXY())
-            # Create mirror
-            mirror = MirrorPoint(v_surf_pts, v_normals, shape, interpolation_type)
-            # Create facet
-            facets.append(Facet(mirror))
-            # Get facet pointing if multi-facet
-            if self.optic_type == 'multi':
-                trans: TransformXYZ = self.data_calculation_ensemble[idx_facet].trans_facet_ensemble
-                trans_list.append(trans)
-
-        # Return either ensemble or facet
-        if self.optic_type == 'multi':
-            ensemble = FacetEnsemble(facets)
-            ensemble.set_facet_transform_list(trans_list)
-            return ensemble
-        else:
-            return facets[0]
